--- core/src/main/java/hello/core/AppConfig.java
@@ -1,27 +1,0 @@
-package hello.core;
-
-import hello.core.discount.RateDiscountPolicy;
-import hello.core.repository.MemoryMemberRepository;
-import hello.core.service.MemberService;
-import hello.core.service.MemberServiceImpl;
-import hello.core.service.OrderService;
-import hello.core.service.OrderServiceImpl;
-
-public class AppConfig {
-
-    public MemberService memberService() {
-        return new MemberServiceImpl(memberRepository());
-    }
-
-    private MemoryMemberRepository memberRepository() {
-        return new MemoryMemberRepository();
-    }
-
-    public OrderService orderService() {
-        return new OrderServiceImpl(memberRepository(), discountPolicy());
-    }
-
-    private RateDiscountPolicy discountPolicy() {
-        return new RateDiscountPolicy();
-    }
-}
