+++ core/src/main/java/personal/core/discount/RateDiscountPolicy.java
@@ -1,7 +1,7 @@
-package hello.core.discount;
+package personal.core.discount;
 
-import hello.core.member.Grade;
-import hello.core.member.Member;
+import personal.core.member.Grade;
+import personal.core.member.Member;
 
 public class RateDiscountPolicy implements DiscountPolicy {
