+++ core/src/main/java/personal/core/member/Grade.java
@@ -1,4 +1,4 @@
-package hello.core.member;
+package personal.core.member;
 
 public enum Grade {
