+++ core/src/main/java/personal/core/member/Member.java
@@ -1,4 +1,4 @@
-package hello.core.member;
+package personal.core.member;
 
 public class Member {
